// client/firebase/types/PastTradeType.ts
export interface PastTradeType {
  tradeUuid: string;
  sendUserUuid: string;
  sendUsername: string;
  sendUserPhoto: string;
  receiveUserUuid: string;
  receiveUsername: string;
  receiveUserPhoto: string;
  sendPinUuid: string;
  sendPinSrc: string;
  receivePinUuid: string;
  receivePinSrc: string;
  date: string;
}